+++ modules/config.js
@@ -1,9 +1,33 @@
 export const ageSystem = {
-    abilitiesSettings: {}
+    abilitiesSettings: {
+        // Ability set for "main" - core AGE System games
+        main: {
+            "acc": "age-system.acc",
+            "comm": "age-system.comm",
+            "cons": "age-system.cons",
+            "dex": "age-system.dex",
+            "fight": "age-system.fight",
+            "int": "age-system.int",
+            "per": "age-system.per",
+            "str": "age-system.str",
+            "will": "age-system.will",
+        },
+        // Ability set for "dage" - Dragon Age games
+        dage: {
+            "comm": "age-system.comm",
+            "cons": "age-system.cons",
+            "cunn": "age-system.cunn",
+            "dex": "age-system.dex",
+            "magic": "age-system.magic",
+            "per": "age-system.per",
+            "str": "age-system.str",
+            "will": "age-system.will",
+        }
+    }
 };
 
-// Ability set for "main" - core AGE System games
-ageSystem.abilitiesSettings.main = {
+// All possible Abilities
+ageSystem.abilitiesTotal = {
     "acc": "age-system.acc",
     "comm": "age-system.comm",
     "cons": "age-system.cons",
@@ -13,19 +37,11 @@ ageSystem.abilitiesSettings.main = {
     "per": "age-system.per",
     "str": "age-system.str",
     "will": "age-system.will",
-};
 
-// Ability set for "dage" - Dragon Age games
-ageSystem.abilitiesSettings.dage = {
-    "comm": "age-system.comm",
-    "cons": "age-system.cons",
     "cunn": "age-system.cunn",
     "dex": "age-system.dex",
-    "magic": "age-system.magic",
-    "per": "age-system.per",
-    "str": "age-system.str",
-    "will": "age-system.will",
-};
+    "magic": "age-system.magic"    
+}
 
 // Organization Abilities
 ageSystem.abilitiesOrg = {
@@ -254,13 +270,71 @@ ageSystem.uiElements = {
 
 const advIconPath =  "systems/age-system/resources/imgs/adv-icon/";
 ageSystem.advIcon = {
-    health: `${advIconPath}heart-key.svg`,
-    ability: `${advIconPath}orb-direction.svg`,
-    newItem: `${advIconPath}upgrade.svg`,
-    progressItem: `${advIconPath}progression.svg`,
-    defense: `${advIconPath}armor-upgrade.svg`,
-    toughness: `${advIconPath}strong.svg`
+    // health: `${advIconPath}heart-key.svg`,
+    // ability: `${advIconPath}orb-direction.svg`,
+    item: `${advIconPath}family-tree.svg`,
+    progressive: `${advIconPath}progression.svg`,
 }
+
+// Advancement Data - progressive values
+ageSystem.adv = {
+    type: {
+        progressive: ['health', 'conviction', 'advAbility', 'defenseAndToughness'/*, 'toughness', 'defense'*/],
+        item: ['spec', 'stunts', 'talent', 'power', 'focus', 'relationship']
+    },
+    health: new Array(20).fill('1 + @cons'),
+    conviction: [1, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1],
+    advAbility: new Array(20).fill(1),
+    defenseAndToughness: [0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1],
+}
+
+/**
+ * Constains path of deverived data, which Active Effects shall be applied after DerivedaData method
+ */
+ageSystem.actorDerivedDataKey = [
+    'system.armor.penaly',
+    'system.armor.strain',
+    'system.dmgMod',
+    'system.testMod',
+    'system.attackMod',
+    // 'system.defense.mod',
+    'system.defense.total',
+    'system.armor.impact',
+    'system.armor.ballistic',
+    // 'system.armor.toughness.mod',
+    'system.armor.toughness.total',
+    // 'system.speed.mod',
+    'system.speed.total',
+    'system.defend.mod',
+    'system.guardUp.mod',
+    'system.allOutAttack.mod',
+    'system.allOutAttack.dmgBonus',
+    // 'system.health.mod',
+    'system.health.max',
+    // 'system.conviction.mod',
+    'system.conviction.max',
+    // 'system.powerPoints.mod',
+    'system.powerPoints.max',
+    'system.aim.mod',
+    'system.initiative',
+    'system.resources.total',
+];
+
+// Ability paths
+ageSystem.charAblKey = [
+    'system.abilities.acc.total',
+    'system.abilities.comm.total',
+    'system.abilities.cons.total',
+    'system.abilities.cunn.total',
+    'system.abilities.dex.total',
+    'system.abilities.fight.total',
+    'system.abilities.int.total',
+    'system.abilities.magic.total',
+    'system.abilities.per.total',
+    'system.abilities.str.total',
+    'system.abilities.will.total',
+];
+
 const AGEstatusEffectsPath = "systems/age-system/resources/imgs/effects/";
 ageSystem.statusEffectsPath = AGEstatusEffectsPath;