+++ modules/dice.js
@@ -293,7 +293,7 @@ export async function ageRollCheck({event = null, actor = null, abl = null, item
         };
     }
 
-    if (!flavor) flavor = actor?.name ?? game.user.name;
+    if (!flavor) flavor = (isToken ? actor?.token.name : actor?.name) ?? game.user.name;
     const stuntFlavor = game.i18n.localize("age-system.stuntAttack");
     switch (rollType) {
         case ROLL_TYPE.ATTACK || ROLL_TYPE.MELEE_ATTACK || RANGED_ATTACK:
@@ -464,6 +464,7 @@ async function getAgeRollOptions(itemRolled, data = {}) {
             buttons: {
                 normal: {
                     label: game.i18n.localize("age-system.roll"),
+                    icon: `<i class="fa-light fa-dice"></i>`,
                     callback: html => {
                         const fd = new FormDataExtended(html[0].querySelector("form"));
                         resolve(fd.object)
@@ -471,6 +472,7 @@ async function getAgeRollOptions(itemRolled, data = {}) {
                 },
                 cancel: {
                     label: game.i18n.localize("age-system.cancel"),
+                    icon: `<i class="fa-solid fa-xmark"></i>`,
                     callback: html => resolve({cancelled: true}),
                 }
             },
@@ -498,6 +500,7 @@ async function getDamageRollOptions(addFocus, stuntDmg, data = {}) {
             buttons: {
                 normal: {
                     label: game.i18n.localize("age-system.roll"),
+                    icon: `<i class="fa-light fa-dice"></i>`,
                     callback: html => {
                         const fd = new FormDataExtended(html[0].querySelector("form"));
                         resolve(fd.object);
@@ -505,6 +508,7 @@ async function getDamageRollOptions(addFocus, stuntDmg, data = {}) {
                 },
                 cancel: {
                     label: game.i18n.localize("age-system.cancel"),
+                    icon: `<i class="fa-solid fa-xmark"></i>`,
                     callback: html => resolve({cancelled: true}),
                 }
             },
@@ -737,7 +741,7 @@ export async function plotDamage (actor) {
         finalValue: dmgRoll.total,
         diceTerms: dmgRoll.terms,
         colorScheme: `colorset-${game.settings.get("age-system", "colorScheme")}`,
-        flavor: actor.name,
+        flavor: actor.isToken ? actor.token.name : actor.name,
         flavor2: "structure damage",
         user: game.user,
         useInjury: undefined, //modificado
@@ -796,6 +800,7 @@ export async function itemDamage({
     actorWgroups = []}={}) {
 
     // Prompt user for Damage Options if Alt + Click is used to initialize damage roll
+    const ownerName = item.actor.isToken ? item.actor.token.name : item.actor.name;
     let damageOptions = null;
     if ((!event.ctrlKey && event.altKey) || event.type === "contextmenu") {
         damageOptions = await getDamageRollOptions(addFocus, stuntDie);
@@ -947,7 +952,7 @@ export async function itemDamage({
         finalValue: wGroupPenalty? Math.floor(dmgRoll.total/2) : dmgRoll.total,
         diceTerms: dmgRoll.terms,
         colorScheme: `colorset-${game.settings.get("age-system", "colorScheme")}`,
-        flavor: item ? `${item.name} | ${item.actor.name}` : damageDesc,
+        flavor: item ? `${item.name} | ${ownerName}` : damageDesc,
         flavor2: item ? damageDesc : null,
         user: game.user,
         useInjury: healthSys.useInjury